refactor: scene uses ecs for objects
Previously the scene defined an object type,
and all objects to be rendered had to present as
that type. Now the scene uses the recently
developed "ecs" module to manage objects. This
has the benefit of allowing certain objects to
define additional components that, when present,
will allow them to be rendered differently than base
objects.