feat: depth testing for renderpasses
Provides an option to enable depth testing for a
renderpass. The API is not as clean as it could be;
for example it is up to the user to make sure depth
testing is enabled for both the renderpass and
the pipeline being bound during it. This is being put
off until we do Vulkan dynamic rendering.