feat: scene object list
The left panel is now populated with the names
of the objects in the scene. As a bonus, selected
an object from that list will bring up an interface
to edit the translation (in the bottom panel).
Doing that required now storing the components
of a transformation (e.g translation) and
reconstructing the matrix when one is changed.
This will be more useful when we have other
transformation components like rotation.