feat/bug: use vulkan dynamic rendering
Modern vulkan supports dynamic rendering, which
can, in some cases, remove the need for constructing
renderpasses and framebuffers. This change
implements it for the offscreen renderpass, though
it's not as generic and sound as it could be.
Since these issues, such as bad rendergraph
compilation and "tagged" renderpasses, are beyond
the scope of dynamic rendering, they'll be
addressed in their own commits. Bug: as a result of
how the rendergraph is poorly designed, resizing
doesn't work since it stores pointers to vulkan
images which are recreated on resize.