Add `Reactable` types
For data such as window size, there is need for several resources
to react to changes to the data. This change provides a generic
type that wraps such data. In doing so there is a single interface
which users can use to register callbacks that will be called when
the data changes. It also encodes in the type system the intended
use and associated overhead of such data.