First version of Vulkan rendergraph
The first iteration of the Vulkan rendergraph is complete.
In order to get a program to compile with it, numerous refactors
were required.

- A resource system now manages most Vulkan resources.
- Commands are no longer driven through a virtual command buffer.
Users provide their own callbacks, which call ordinary Vulkan
commands.

A number of features are now broken.

- Reverting back to a simpler render. Only the basic "hello
triangle" was tested. This includes the removal of ImGUI
support.
- Resizing. The program won't crash, but it won't look right.