Fix window resizing
Window resizing works and the image is as expected. With the
addition of `Reactable` types this solution is more idiomatic
than it ever was before. One issue to be weary of, which caused
trouble while debugging, is that Vulkan almost always expects
pixel dimensions rather than device-independent dimensions
(the two may differ due to DPI).

(fixes #15)