Clean up pipeline creation
Aggressive refactoring left pipeline creation a mess, from
functions that were not generic enough and reused a lot of code
to functions completely seperate from the pipeline system.
This change makes the pipeline system a lot simpler-- it is just
responsible for storing pipelines and pipeline layouts, and does
not have anything to do with caching or virtual pipelines.