Enable building multiple examples through CLI
Currently the file being compiled is hard-coded into the
`build.zig` file. With this change we introduce the ability to
add multiple examples to build. On the CLI, a flag is generated
which, for each example specified, will build that example.
There is also a flag to build all examples.

(fixes #16)