Support offscreen rendering, with example
There is a new example demonstrating rendering a triangle to
an image and displaying that image through ImGui. Along the way
there was a change in the rendergraph semantics-- previously you
would list "final" resources as inputs and outputs, but now it
should only be listed as an output. The thinking is that "inputs"
are reserved for resources the renderpass accesses during
rendering.